+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 
 - Changed the refreshable status check to always conside elements with `live` and `active` statuses as refreshable ([#572](https://github.com/putyourlightson/craft-blitz/issues/572)).
+- Refreshing expired cached now forcibly generates new cached pages if they are not cleared ([#573](https://github.com/putyourlightson/craft-blitz/issues/573)).
 
 ### Fixed